CRecursiveMutex: move to separate file
